Move typescript types into package
