+++ typescript/schemas/src/jsonschema.ts
@@ -1,1 +1,3 @@
-export * from "../../../schemas/jsonschema";
+// This file only exists to ensure the jsonschema directory was copied.
+// If you get an build error you need to run `yarn prepack` first.
+export * from "./jsonschema/index.js";
