+++ typescript/schemas/src/types/CameraCalibration.ts
