+++ typescript/schemas/src/types/CircleAnnotation.ts
