+++ typescript/schemas/src/types/Color.ts
