+++ typescript/schemas/src/types/CompressedImage.ts
