+++ typescript/schemas/src/types/CompressedVideo.ts
