+++ typescript/schemas/src/types/CubePrimitive.ts
