+++ typescript/schemas/src/types/ImageAnnotations.ts
