+++ typescript/schemas/src/types/LaserScan.ts
