+++ typescript/schemas/src/types/Point2.ts
