+++ typescript/schemas/src/types/Point3.ts
