+++ typescript/schemas/src/types/PointCloud.ts
