+++ typescript/schemas/src/types/PointsAnnotation.ts
