+++ typescript/schemas/src/types/PointsAnnotationType.ts
