+++ typescript/schemas/src/types/Pose.ts
