+++ typescript/schemas/src/types/PoseInFrame.ts
