+++ typescript/schemas/src/types/PosesInFrame.ts
