+++ typescript/schemas/src/types/Quaternion.ts
