+++ typescript/schemas/src/types/RawImage.ts
