+++ typescript/schemas/src/types/SceneEntity.ts
