+++ typescript/schemas/src/types/SceneEntityDeletion.ts
