+++ typescript/schemas/src/types/SceneEntityDeletionType.ts
