+++ typescript/schemas/src/types/SceneUpdate.ts
