+++ typescript/schemas/src/types/SpherePrimitive.ts
