+++ typescript/schemas/src/types/TextAnnotation.ts
