+++ typescript/schemas/src/types/TriangleListPrimitive.ts
