+++ typescript/schemas/src/types/Vector3.ts
