+++ typescript/schemas/tsconfig.json
@@ -2,6 +2,7 @@
   "extends": "@foxglove/tsconfig/base",
   "include": ["./**/*"],
   "compilerOptions": {
+    "rootDir": "./src",
     "outDir": "./dist",
     "module": "commonjs",
     "lib": ["es2020"],
